usm: sharedlibraries: tests: Move createTestFile to ebpf_test.go
Moving the method to the ebpf_test.go, as we're about to delete watcher_test.go
and the method is still needed
